requirements.txt
pytest
pytest-html
pytest-xdist
pytest-ordering
pytest-rerunfailures
pytest-base-url
allure-pytest
pyyaml
requests
jsonpath
pymysql
mysqlclient
rsa